+++ apps/web/src/lib/components/ConfirmationModal.svelte
@@ -24,6 +24,7 @@
 		<h1 class="text-sm my-3">{message}</h1>
 		<div class="flex flex-row justify-center">
 			<button
+				type="button"
 				on:click={handleClose}
 				class="p-1 m-3 px-3 border-2 text-center border-slate-200 shadow-md rounded"
                 data-testid="close"
@@ -32,6 +33,7 @@
 			</button>
 
 			<button
+				type="submit"
 				on:click={handleConfirm}
 				class="p-1 m-3 px-3 border-2 text-center border-red-800 bg-red-500 text-white shadow-md rounded"
                 data-testid="confirm"
