+++ apps/web/src/routes/(app)/home/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import LinkButton from '$lib/components/LinkButton.svelte';
-	import { routes } from '$lib/types/routes';
+	import { routes } from '$lib/const/routes';
 </script>
 
 <div class="flex flex-col items-center mx-5">
