+++ apps/web/src/routes/(app)/menus/+page.server.ts
@@ -1,10 +1,9 @@
-import type { PageServerLoad } from './$types';
+import type { Actions, PageServerLoad } from './$types';
 import { redirectIfNoSession } from '$lib/server/auth';
-import { error, type Actions } from '@sveltejs/kit';
+import { error, fail } from '@sveltejs/kit';
+import { getStringFromFormValue } from '$lib/forms/form-input';
 
-export const load: PageServerLoad = async ({ parent, locals, depends }) => {
-	depends('menus:load');
-	console.log('/menus load function');
+export const load: PageServerLoad = async ({ parent, locals }) => {
 	await parent();
 	const { session, user, supabase } = locals;
 	redirectIfNoSession(session);
@@ -18,20 +17,75 @@ export const load: PageServerLoad = async ({ parent, locals, depends }) => {
 	if (supabaseError) {
 		error(500, { message: supabaseError.message, name: 'DBError' });
 	}
-	console.log('data', data);
 	return { menus: data };
 };
 
-export const actions: Actions = {
+export const actions = {
 	createMenu: async ({ locals }) => {
 		const { session, user, supabase } = locals;
 
-		redirectIfNoSession(session);
+		if (!session || !user) {
+			error(401);
+		}
 
 		if (!user) {
 			error(500, { message: 'no user', name: 'NoUserError' });
 		}
 
-		await supabase.from('menus').insert({ name: 'My New Menu', user_id: user.id });
+		const { error: pgError } = await supabase
+			.from('menus')
+			.insert({ name: 'My New Menu', user_id: user.id });
+
+		if (pgError) {
+			fail(500, { message: pgError.message });
+		}
+	},
+	renameMenu: async ({ locals, request }) => {
+		const { session, user, supabase } = locals;
+
+		if (!session || !user) {
+			error(401);
+		}
+
+		const formData = await request.formData();
+		const name = getStringFromFormValue(formData.get('name'));
+		const menuId = getStringFromFormValue(formData.get('id'));
+
+		if (!name || !null) {
+			fail(400, { message: 'Missing inputs', name });
+		}
+
+		const { error: pgError } = await supabase
+			.from('menus')
+			.update({ name: name })
+			.eq('id', menuId)
+			.eq('user_id', user.id)
+			.select();
+
+		if (pgError) {
+			fail(500, { message: pgError.message, name });
+		}
+
+		return { name, success: true };
+	},
+	deleteMenu: async ({ locals, request }) => {
+		const { session, user, supabase } = locals;
+
+		if (!session || !user) {
+			error(401);
+		}
+
+		const formData = await request.formData();
+		const menuId = getStringFromFormValue(formData.get('id'));
+
+		const { error: pgError } = await supabase
+			.from('menus')
+			.delete()
+			.eq('id', menuId)
+			.eq('user_id', user.id);
+
+		if (pgError) {
+			error(500, { message: pgError.message, name: 'DbError' });
+		}
 	}
-};
+} satisfies Actions;
