+++ apps/web/src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Environment from '$lib/components/Environment.svelte';
 	import LinkButton from '$lib/components/LinkButton.svelte';
-	import { routes } from '$lib/types/routes';
+	import { routes } from '$lib/const/routes';
 </script>
 
 <div class="flex flex-col items-center mx-5">
