+++ apps/web/src/routes/auth/signup/pending/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import LinkButton from '$lib/components/LinkButton.svelte';
-	import { routes } from '$lib/types/routes.js';
+	import { routes } from '$lib/const/routes.js';
 	export let data;
 </script>
