+++ apps/web/src/routes/error/not-signed-in/+page.svelte
@@ -1,6 +1,6 @@
 <script>
 	import LinkButton from '$lib/components/LinkButton.svelte';
-	import { routes } from '$lib/types/routes';
+	import { routes } from '$lib/const/routes';
 </script>
 
 <div class="flex flex-col items-center">
